rube-deployment/requirements.txt
# Requirements for Rube.app deployment
composio-core>=0.3.0
composio-rube>=0.2.0
python-dotenv>=1.0.0
asyncio-throttle>=1.0.0
pydantic>=2.0.0
httpx>=0.24.0